// sanity.types.ts
export type Cv_entry = {
  _id: string;
  _type: "cv_entry";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  order?: number;
  title?: string;
  content?: Array<{
    children?: Array<{
      marks?: Array<string>;
      text?: string;
      _type: "span";
      _key: string;
    }>;
    style?: "normal" | "h1" | "h2" | "h3" | "h4" | "h5" | "h6" | "blockquote";
    listItem?: "bullet" | "number";
    markDefs?: Array<{
      href?: string;
      _type: "link";
      _key: string;
    }>;
    level?: number;
    _type: "block";
    _key: string;
  }>;
};

export type Paper = {
  _id: string;
  _type: "paper";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  order?: number;
  title?: string;
  journal?: string;
  year?: number;
  issueAndPages?: string;
  abstract?: Array<{
    children?: Array<{
      marks?: Array<string>;
      text?: string;
      _type: "span";
      _key: string;
    }>;
    style?: "normal" | "h1" | "h2" | "h3" | "h4" | "h5" | "h6" | "blockquote";
    listItem?: "bullet" | "number";
    markDefs?: Array<{
      href?: string;
      _type: "link";
      _key: string;
    }>;
    level?: number;
    _type: "block";
    _key: string;
  }>;
  cover?: {
    _ref: string;
    _type: "reference";
    _weak?: boolean;
    [internalGroqTypeReferenceTo]?: "picture";
  };
  draftAvailable?: boolean;
  published?: boolean;
  openAccess?: boolean;
  link?: string;
  draft_pdf?: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.fileAsset";
    };
    _type: "file";
  };
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;